Move DB models back to their original package
