+++ src/main/java/uk/gov/hmcts/reform/bulkscanprocessor/entity/DbPayment.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.bulkscanprocessor.model.db;
+package uk.gov.hmcts.reform.bulkscanprocessor.entity;
 
 import java.math.BigDecimal;
 import java.util.UUID;
