+++ src/main/java/uk/gov/hmcts/reform/bulkscanprocessor/entity/ProcessEvent.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.bulkscanprocessor.model.db;
+package uk.gov.hmcts.reform.bulkscanprocessor.entity;
 
 import uk.gov.hmcts.reform.bulkscanprocessor.model.common.Event;
