+++ src/main/java/uk/gov/hmcts/reform/bulkscanprocessor/model/in/StatusUpdate.java
@@ -1,7 +1,7 @@
 package uk.gov.hmcts.reform.bulkscanprocessor.model.in;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import uk.gov.hmcts.reform.bulkscanprocessor.model.common.Status;
+import uk.gov.hmcts.reform.bulkscanprocessor.entity.Status;
 
 public class StatusUpdate {
