+++ src/test/java/uk/gov/hmcts/reform/bulkscanprocessor/model/mapper/EnvelopeResponseMapperTest.java
@@ -2,10 +2,10 @@ package uk.gov.hmcts.reform.bulkscanprocessor.model.mapper;
 
 import org.junit.Before;
 import org.junit.Test;
+import uk.gov.hmcts.reform.bulkscanprocessor.entity.DbEnvelope;
 import uk.gov.hmcts.reform.bulkscanprocessor.helper.EnvelopeCreator;
 import uk.gov.hmcts.reform.bulkscanprocessor.helper.ToStringComparator;
 import uk.gov.hmcts.reform.bulkscanprocessor.model.common.Classification;
-import uk.gov.hmcts.reform.bulkscanprocessor.model.db.DbEnvelope;
 import uk.gov.hmcts.reform.bulkscanprocessor.model.out.EnvelopeResponse;
 
 import static org.assertj.core.api.Assertions.assertThat;
