+++ src/test/java/uk/gov/hmcts/reform/bulkscanprocessor/services/EnvelopeStatusChangeValidatorTest.java
@@ -1,14 +1,14 @@
 package uk.gov.hmcts.reform.bulkscanprocessor.services;
 
 import org.junit.Test;
+import uk.gov.hmcts.reform.bulkscanprocessor.entity.Status;
 import uk.gov.hmcts.reform.bulkscanprocessor.exceptions.InvalidStatusChangeException;
-import uk.gov.hmcts.reform.bulkscanprocessor.model.common.Status;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static uk.gov.hmcts.reform.bulkscanprocessor.model.common.Status.CONSUMED;
-import static uk.gov.hmcts.reform.bulkscanprocessor.model.common.Status.NOTIFICATION_SENT;
-import static uk.gov.hmcts.reform.bulkscanprocessor.model.common.Status.UPLOAD_FAILURE;
+import static uk.gov.hmcts.reform.bulkscanprocessor.entity.Status.CONSUMED;
+import static uk.gov.hmcts.reform.bulkscanprocessor.entity.Status.NOTIFICATION_SENT;
+import static uk.gov.hmcts.reform.bulkscanprocessor.entity.Status.UPLOAD_FAILURE;
 
 public class EnvelopeStatusChangeValidatorTest {
