+++ src/test/java/uk/gov/hmcts/reform/bulkscanprocessor/services/EnvelopeUpdateServiceTest.java
@@ -6,13 +6,13 @@ import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
+import uk.gov.hmcts.reform.bulkscanprocessor.entity.DbEnvelope;
 import uk.gov.hmcts.reform.bulkscanprocessor.entity.EnvelopeRepository;
+import uk.gov.hmcts.reform.bulkscanprocessor.entity.ProcessEvent;
 import uk.gov.hmcts.reform.bulkscanprocessor.entity.ProcessEventRepository;
 import uk.gov.hmcts.reform.bulkscanprocessor.exceptions.EnvelopeNotFoundException;
 import uk.gov.hmcts.reform.bulkscanprocessor.helper.EnvelopeCreator;
 import uk.gov.hmcts.reform.bulkscanprocessor.model.common.Event;
-import uk.gov.hmcts.reform.bulkscanprocessor.model.db.DbEnvelope;
-import uk.gov.hmcts.reform.bulkscanprocessor.model.db.ProcessEvent;
 
 import java.util.Optional;
 import java.util.UUID;
@@ -24,17 +24,21 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
-import static uk.gov.hmcts.reform.bulkscanprocessor.model.common.Status.CONSUMED;
-import static uk.gov.hmcts.reform.bulkscanprocessor.model.common.Status.UPLOAD_FAILURE;
+import static uk.gov.hmcts.reform.bulkscanprocessor.entity.Status.CONSUMED;
+import static uk.gov.hmcts.reform.bulkscanprocessor.entity.Status.UPLOAD_FAILURE;
 
 @SuppressWarnings({"PMD.BeanMembersShouldSerialize", "checkstyle:linelength"})
 @RunWith(MockitoJUnitRunner.class)
 public class EnvelopeUpdateServiceTest {
 
-    @Mock private EnvelopeRepository envelopeRepo;
-    @Mock private ProcessEventRepository eventRepo;
-    @Mock private EnvelopeAccessService accessService;
-    @Mock private EnvelopeStatusChangeValidator statusChangeValidator;
+    @Mock
+    private EnvelopeRepository envelopeRepo;
+    @Mock
+    private ProcessEventRepository eventRepo;
+    @Mock
+    private EnvelopeAccessService accessService;
+    @Mock
+    private EnvelopeStatusChangeValidator statusChangeValidator;
 
     private EnvelopeUpdateService service;
