+++ src/test/java/uk/gov/hmcts/reform/bulkscanprocessor/tasks/SendNotificationTaskTest.java
@@ -7,8 +7,8 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 import uk.gov.hmcts.reform.bulkscanprocessor.entity.EnvelopeRepository;
 import uk.gov.hmcts.reform.bulkscanprocessor.entity.ProcessEventRepository;
+import uk.gov.hmcts.reform.bulkscanprocessor.entity.Status;
 import uk.gov.hmcts.reform.bulkscanprocessor.exceptions.InvalidMessageException;
-import uk.gov.hmcts.reform.bulkscanprocessor.model.common.Status;
 import uk.gov.hmcts.reform.bulkscanprocessor.services.servicebus.ServiceBusHelper;
 
 import static java.util.stream.Collectors.toList;
@@ -23,9 +23,12 @@ import static uk.gov.hmcts.reform.bulkscanprocessor.helper.EnvelopeCreator.envel
 @RunWith(MockitoJUnitRunner.class)
 public class SendNotificationTaskTest {
 
-    @Mock private ServiceBusHelper serviceBusHelper;
-    @Mock private EnvelopeRepository envelopeRepo;
-    @Mock private ProcessEventRepository processEventRepo;
+    @Mock
+    private ServiceBusHelper serviceBusHelper;
+    @Mock
+    private EnvelopeRepository envelopeRepo;
+    @Mock
+    private ProcessEventRepository processEventRepo;
 
     private SendNotificationTask task;
